Factor out common html page lib
