+++ lib/ui/cardano-wallet-ui.cabal
@@ -40,6 +40,7 @@ library
     Cardano.Wallet.UI.Common.Html.Html
     Cardano.Wallet.UI.Common.Html.Htmx
     Cardano.Wallet.UI.Common.Html.Lib
+    Cardano.Wallet.UI.Common.Html.Pages.Lib
     Cardano.Wallet.UI.Common.Html.Pages.Template.Footer
     Cardano.Wallet.UI.Common.Html.Pages.Template.Head
     Cardano.Wallet.UI.Common.Html.Pages.Template.Navigation
@@ -52,7 +53,6 @@ library
     Cardano.Wallet.UI.Personal.Handlers.Wallets
     Cardano.Wallet.UI.Personal.Html.Pages.About
     Cardano.Wallet.UI.Personal.Html.Pages.Addresses
-    Cardano.Wallet.UI.Personal.Html.Pages.Lib
     Cardano.Wallet.UI.Personal.Html.Pages.Network
     Cardano.Wallet.UI.Personal.Html.Pages.Page
     Cardano.Wallet.UI.Personal.Html.Pages.Settings
